Change option type to work better on multiple platforms

+++ lib/Option.cpp
@@ -123,7 +123,7 @@ llvm::cl::opt<bool> constant_args_in_uniform_buffer(
     llvm::cl::desc("Put pointer-to-constant kernel args in UBOs."));
 
 // Default to 64kB.
-llvm::cl::opt<uint64_t> maximum_ubo_size(
+llvm::cl::opt<int> maximum_ubo_size(
     "max-ubo-size", llvm::cl::init(64 << 10),
     llvm::cl::desc("Specify the maximum UBO array size in bytes."));
